requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
python-dotenv==1.0.1
streamlit==1.39.0
seaborn==0.13.2
plotly==5.24.1
